requirements.txt
PyQt5==5.15.10
PyQt5-sip==12.13.0
hdwallet==2.2.1
bip39==0.0.1
web3==6.11.3
bitcoinlib==0.6.14
cryptography==41.0.7
qrcode==7.4.2
pillow==9.5.0 